3lab: refactor implementation and create template for CLI and GUI
